re-compose modal
