+++ ui/app/routes/observability/episodes/$episode_id/route.tsx
@@ -12,6 +12,7 @@ import {
 import type { Route } from "./+types/route";
 import {
   data,
+  Form,
   isRouteErrorResponse,
   redirect,
   useNavigate,
@@ -26,11 +27,17 @@ import {
   SectionsGroup,
   SectionHeader,
 } from "~/components/layout/PageLayout";
-import EpisodeActions from "./EpisodeActions";
 import { addHumanFeedback } from "~/utils/tensorzero.server";
 import { Toaster } from "~/components/ui/toaster";
 import { useToast } from "~/hooks/use-toast";
-import { useEffect } from "react";
+import { useEffect, useState } from "react";
+import { ActionBar } from "~/components/layout/ActionBar";
+import { HumanFeedbackButton } from "~/components/feedback/HumanFeedbackButton";
+import { HumanFeedbackModal } from "~/components/feedback/HumanFeedbackModal";
+import { HumanFeedbackForm } from "~/components/feedback/HumanFeedbackForm";
+
+const FF_ENABLE_FEEDBACK =
+  import.meta.env.VITE_TENSORZERO_UI_FF_ENABLE_FEEDBACK === "1";
 
 export async function loader({ request, params }: Route.LoaderArgs) {
   const { episode_id } = params;
@@ -126,6 +133,7 @@ export default function InferencesPage({ loaderData }: Route.ComponentProps) {
     newFeedbackId,
   } = loaderData;
   const navigate = useNavigate();
+  const [isModalOpen, setIsModalOpen] = useState(false);
 
   const topInference = inferences[0];
   const bottomInference = inferences[inferences.length - 1];
@@ -191,7 +199,20 @@ export default function InferencesPage({ loaderData }: Route.ComponentProps) {
   return (
     <PageLayout>
       <PageHeader label="Episode" name={episode_id}>
-        <EpisodeActions episodeId={episode_id} />
+        {FF_ENABLE_FEEDBACK && (
+          <ActionBar>
+            <HumanFeedbackModal
+              isOpen={isModalOpen}
+              onOpenChange={setIsModalOpen}
+              trigger={<HumanFeedbackButton />}
+            >
+              <Form method="post" onSubmit={() => setIsModalOpen(false)}>
+                <input type="hidden" name="_action" value="addFeedback" />
+                <HumanFeedbackForm episodeId={episode_id} />
+              </Form>
+            </HumanFeedbackModal>
+          </ActionBar>
+        )}
       </PageHeader>
 
       <SectionsGroup>
